// api-client/models/create-category-dto.ts
/* tslint:disable */
 
/**
 * API Documentation
 * API - IT Asset Tracker
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface CreateCategoryDto
 */
export interface CreateCategoryDto {
    /**
     * Name of the category
     * @type {string}
     * @memberof CreateCategoryDto
     */
    'name': string;
    /**
     * Unique category code.
     * @type {string}
     * @memberof CreateCategoryDto
     */
    'code': string;
    /**
     * Description of the category.
     * @type {string}
     * @memberof CreateCategoryDto
     */
    'description': string;
}